% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main_Measure_Bias.R
\name{exe}
\alias{exe}
\title{exe()}
\usage{
exe()
}
\arguments{
\item{none}{}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{   A folder with analysis results
}\if{html}{\out{</div>}}
}
\description{
This function performs complete data analysis. This function coordinates
the loading, cleaning, analysis and visualization of the data.
}
\examples{
       exe()
}
