% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Measure_Bias.R
\name{F_Adjusted_histogram}
\alias{F_Adjusted_histogram}
\title{F_Adjusted_histogram}
\usage{
F_Adjusted_histogram(Dataframe, num_classes)
}
\arguments{
\item{Dataframe}{: Data to generate the plot
num_classes : Number of classes or partitions for the plot}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{   Adjusted scatter plot (range 50 - 100)
}\if{html}{\out{</div>}}
}
\description{
Function to create adjusted histogram
}
\examples{
       Plots <- F_Adjusted_histogram(Dataframe,num_classes)
}
