% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Measure_Bias.R
\name{F_Correlation_graph}
\alias{F_Correlation_graph}
\title{F_Correlation_graph}
\usage{
F_Correlation_graph(Dataframe)
}
\arguments{
\item{A}{dataframe with the 2 variables of interest}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{   A PGN correlation graph
}\if{html}{\out{</div>}}
}
\description{
Function to create a correlation graph between two variables in a given format.
}
\examples{
       plot <- F_Correlation_graph(Dataframe)
}
