% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Measure_Bias.R
\name{F_Create_outputs_folder}
\alias{F_Create_outputs_folder}
\title{F_Create_outputs_folder}
\usage{
F_Create_outputs_folder(Outputs)
}
\arguments{
\item{Outputs:}{A name for the output folder}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{   None
}\if{html}{\out{</div>}}
}
\description{
Function to create folder and subfolder for outputs
}
\examples{
       F_Create_outputs_folder(Outputs)
}
