% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Measure_Bias.R
\name{F_Data_information}
\alias{F_Data_information}
\title{F_Data_information}
\usage{
F_Data_information(
  Bias_data,
  Population_data,
  Active_population,
  Boundaries,
  Bias_NA,
  Bias_NEG
)
}
\arguments{
\item{Bias_data}{: data with bias
Population_data  : data from population like data census
Active_population: data from other source like facebook or phone
Boundaries       : data from shp file
Bias_NA          : dataframe with only NA values
Bias_NEG         : dataframe with only Negative bias}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{   A dataframe with the dataset information
}\if{html}{\out{</div>}}
}
\description{
Function to summarise dataset information
}
\examples{
       Data <- F_Drop_negative_bias(Bias,Population,Active,Boundaries,NA,Negatives)
}
