% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Measure_Bias.R
\name{F_Drop_negative_bias}
\alias{F_Drop_negative_bias}
\title{F_Drop_negative_bias}
\usage{
F_Drop_negative_bias(Dataframe)
}
\arguments{
\item{Data:}{dataframe without negative biases}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{   A dataframe without negative biases and a dataframe only with
   negative biases
}\if{html}{\out{</div>}}
}
\description{
Function to eliminate negative biases
}
\examples{
       Data <- F_Drop_negative_bias(Dataframe)
}
