% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Measure_Bias.R
\name{F_Histograms}
\alias{F_Histograms}
\title{F_Histograms}
\usage{
F_Histograms(Dataframe, num_classes)
}
\arguments{
\item{Dataframe}{: Data to generate the plot
num_classes : Number of classes or partitions for the plot}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{   2 histograms:Labelled_histogram and Unlabelled_histogram
}\if{html}{\out{</div>}}
}
\description{
Function to create histograms
}
\examples{
       Plots <- F_Histograms(Dataframe,num_classes)
}
