% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Measure_Bias.R
\name{F_Listwise_deletion}
\alias{F_Listwise_deletion}
\title{F_Listwise_deletion}
\usage{
F_Listwise_deletion(Dataframe)
}
\arguments{
\item{Data:}{dataframe with missing values}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{   A dataframe without missing values and a dataframe only with
   missing values
}\if{html}{\out{</div>}}
}
\description{
Function for handling missing values using listwise deletion method
}
\examples{
       Data <- F_Listwise_deletion(Dataframe)
}
