% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Measure_Bias.R
\name{F_Load_join_images}
\alias{F_Load_join_images}
\title{F_Load_join_images}
\usage{
F_Load_join_images(path_01, path_02A, path_02B, path_03)
}
\arguments{
\item{path_01}{: map_path
path_02A: unlabelled_histogram_path
path_02B: adjusted_histogram_path
path_03 : scatter_path}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{   an image with 3 plots
}\if{html}{\out{</div>}}
}
\description{
Function to load and join images
}
\examples{
       Final_plot <- F_Load_join_images(path_01,path_02A,path_02B,path_03)
}
