% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Measure_Bias.R
\name{F_Map}
\alias{F_Map}
\title{F_Map}
\usage{
F_Map(
  Bias_measurement,
  Bias_map,
  num_classes,
  spatial = 0,
  mI = 0,
  p_value_mI = 0
)
}
\arguments{
\item{Bias_measurement}{:data with bias
Bias_map         :data with spatial correlation
num_classes      :number of classes or partitions
spatial          : 0 for nonspatial 1 for spatial
mI               :Moran's Indicator
mp_value_mI      :p-value}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{   Map with bias
}\if{html}{\out{</div>}}
}
\description{
Function to create maps
}
\examples{
       Map <- F_Map(Bias_measurement,Bias_map,num_classes)
       Map <- F_Map(Bias_measurement,Bias_map,num_classes,1, mI, mp_value_mI)
}
