% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Measure_Bias.R
\name{F_Measure_bias}
\alias{F_Measure_bias}
\title{F_Measure_bias}
\usage{
F_Measure_bias(Population_data, Active_population)
}
\arguments{
\item{Population_data:}{data from population like data census
Active_population: data from other source like facebook or phone}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{   A dataset with bias measurement
}\if{html}{\out{</div>}}
}
\description{
Function to obtain the bias between 2 datasets
}
\examples{
       Bias_data <- F_Measure_bias(Dataframe01, Dataframe02)
}
