% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Measure_Bias.R
\name{F_Moran_results}
\alias{F_Moran_results}
\title{F_Moran_results}
\usage{
F_Moran_results(
  w_schemes,
  Structure_queen,
  Structure_fbw,
  Structure_knn,
  Structure_db
)
}
\arguments{
\item{Bias_map}{: Geographical data with bias
Structure_queen
Structure_fbw
Structure_knn
Structure_db}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{   A dataframe with Moran values, p-values, z-score and others
}\if{html}{\out{</div>}}
}
\description{
Function to Compute Moran's I, p_value & Z-score for spatial cases
}
\examples{
       Data <- F_Moran_results(Bias_map)
}
