% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Measure_Bias.R
\name{F_Saving_csv}
\alias{F_Saving_csv}
\title{F_Saving_csv}
\usage{
F_Saving_csv(Outputs, Name, Data)
}
\arguments{
\item{Outputs:}{Folder and path where the file will be stored
Name   : Name under which the file will be stored
Data   : Data or file to be saved}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{   None
}\if{html}{\out{</div>}}
}
\description{
Function to save files in csv format
}
\examples{
       F_Saving_csv(Outputs,Name,Data)
}
