% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Measure_Bias.R
\name{F_Saving_png}
\alias{F_Saving_png}
\title{F_Saving_png}
\usage{
F_Saving_png(Outputs, Name, Plot, Option = 0)
}
\arguments{
\item{Outputs:}{Folder and path where the file will be stored
Name   : Name under which the file will be stored
Plot   : Plot to be saved
Option : 0 - default image measurements / other number - predefined image measurements}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{   None
}\if{html}{\out{</div>}}
}
\description{
Function to save files in png format
}
\examples{
       F_Saving_csv(Outputs,Name,Plot)
}
