% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Measure_Bias.R
\name{F_Scatter_plot}
\alias{F_Scatter_plot}
\title{F_Scatter_plot}
\usage{
F_Scatter_plot(Dataframe, num_classes)
}
\arguments{
\item{Dataframe}{: Data to generate the plot
num_classes : Number of classes or partitions for the plot}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{   Scatter plot
}\if{html}{\out{</div>}}
}
\description{
Function to create a scatter plot
}
\examples{
       Plot <- F_Scatter_plot(Dataframe,num_classes)
}
