% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Measure_Bias.R
\name{F_Searching_optimum_k}
\alias{F_Searching_optimum_k}
\title{F_Searching_optimum_k}
\usage{
F_Searching_optimum_k(Bias_map)
}
\arguments{
\item{Bias_map}{: Geographical data with bias}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{   best_k            :optimal k from the analysis
   knn_figure        :plots of the results of the analysis
   moran_results_df  :file of the result of the analysis
}\if{html}{\out{</div>}}
}
\description{
Function to search the optimum k for knn
}
\examples{
       Knn_analysis <- F_Searching_optimum_k(Bias_map)
}
