% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Measure_Bias.R
\name{F_Spatial_fbw}
\alias{F_Spatial_fbw}
\title{F_Spatial_fbw}
\usage{
F_Spatial_fbw(Bias_map)
}
\arguments{
\item{Bias_map}{: Geographical data with bias}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{   A structure with spatial lag, number of subgraphs and weight matrix
}\if{html}{\out{</div>}}
}
\description{
Function to Compute spatial lag with fbw method
}
\examples{
       Structure <- F_Spatial_queen(Bias_map)
}
