% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Measure_Bias.R
\name{F_Spatial_knn}
\alias{F_Spatial_knn}
\title{F_Spatial_knn}
\usage{
F_Spatial_knn(Bias_map, k)
}
\arguments{
\item{Bias_map}{: Geographical data with bias
k         : Optimal value of k}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{   A structure with spatial lag, number of subgraphs and weight matrix
}\if{html}{\out{</div>}}
}
\description{
Function to Compute spatial lag with knn method
}
\examples{
       Structure <- F_Spatial_knn(Bias_map,k)
}
