% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Measure_Bias.R
\name{theme_map_tufte}
\alias{theme_map_tufte}
\title{theme_map_tufte}
\usage{
theme_map_tufte(...)
}
\arguments{
\item{none}{}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{   A ggplot2 theme object.
}\if{html}{\out{</div>}}
}
\description{
This function applies a custom theme to the maps.
}
\examples{
       theme_map_tufte()
}
